+++ src/app/[lang]/[method]/[[...data]]/page.module.scss
