+++ src/app/[lang]/[method]/[[...data]]/page.tsx
