+++ src/app/[lang]/graphiql/page.tsx
