+++ src/app/[lang]/history/page.module.scss
