+++ src/app/[lang]/layout.tsx
@@ -1,7 +1,7 @@
 import { Montserrat } from 'next/font/google';
 import type { PropsWithChildren } from 'react';
 import type { Metadata } from 'next';
-import '../styles/scss/style.scss';
+import '../../styles/scss/style.scss';
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v13-appRouter';
 import { CssBaseline, ThemeProvider } from '@mui/material';
 import { theme } from '@/theme.ts';
