+++ src/app/[lang]/not-found.module.scss
@@ -13,7 +13,7 @@
   &__title {
     display: inline-flex;
     height: auto;
-    margin-top: 0px;
+    margin-top: 0;
     font-size: 100px;
     font-weight: 700;
     line-height: 1;
