+++ src/app/[lang]/page.module.scss
@@ -5,7 +5,7 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-image: url('../../public/background.jpg');
+  background-image: url('../../../public/background.jpg');
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
