+++ src/app/[lang]/sign-up/page.tsx
