+++ src/components/historyList/historyList.tsx
@@ -8,8 +8,9 @@ import HistoryListItem from '@components/historyListItem/historyListItem.tsx';
 import { ROUTES } from '@/utils/constants/routes.ts';
 import { useEffect, useState } from 'react';
 import { HistoryItem } from '@components/historyList/types.ts';
+import Loader from '@components/loader/loader.tsx';
 
-function HistoryList() {
+function HistoryList(): JSX.Element {
   const [localStorage] = useLocalStorage('history', '{}');
   const [historyList, setHistoryList] = useState<HistoryItem[]>([]);
   const [isLoaded, setIsLoaded] = useState(false);
@@ -20,30 +21,36 @@ function HistoryList() {
     setIsLoaded(true);
   }, [localStorage]);
 
+  if (!isLoaded) {
+    return <Loader />;
+  }
+
   return (
-    isLoaded && (
-      <Paper className={styles.Paper}>
-        {historyList.length > 0 ? (
-          <List className={styles.List}>
-            {historyList
-              .sort((item1, item2) => item1.executedAt - item2.executedAt)
-              .map((item) => (
-                <HistoryListItem key={item.executedAt} item={item} />
-              ))}
-          </List>
-        ) : (
-          <div>
-            <Typography variant={'body1'} className={styles.Paragraph}>
-              History is empty. Make first request:{' '}
-            </Typography>
-            <div className={styles.Nav}>
-              <Link href={ROUTES.RESTFUL_CLIENT_PATH}>RESTful client</Link>
-              <Link href={ROUTES.GRAPHIQL_PATH}>GraphQL</Link>
+    <>
+      {isLoaded && (
+        <Paper className={styles.Paper}>
+          {historyList.length > 0 ? (
+            <List className={styles.List}>
+              {historyList
+                .sort((item1, item2) => item1.executedAt - item2.executedAt)
+                .map((item) => (
+                  <HistoryListItem key={item.executedAt} item={item} />
+                ))}
+            </List>
+          ) : (
+            <div>
+              <Typography variant={'body1'} className={styles.Paragraph}>
+                History is empty. Make first request:{' '}
+              </Typography>
+              <div className={styles.Nav}>
+                <Link href={ROUTES.RESTFUL_CLIENT_PATH}>RESTful client</Link>
+                <Link href={ROUTES.GRAPHIQL_PATH}>GraphQL</Link>
+              </div>
             </div>
-          </div>
-        )}
-      </Paper>
-    )
+          )}
+        </Paper>
+      )}
+    </>
   );
 }
