--- src/components/monacoEditor/monacoEditor.module.scss
@@ -1,17 +1,0 @@
-.Editor {
-  height: calc(100% - 2px);
-}
-
-.Editor section > div {
-  border: 1px solid rgba(81, 81, 81, 1);
-  border-radius: 4px;
-  overflow: hidden;
-
-  &:hover {
-    border-color: rgba(230, 224, 233, 0.8);
-  }
-
-  &:has(textarea:focus) {
-    border-color: #e6e0e9;
-  }
-}
