+++ src/components/responseStatus/responseStatus.tsx
@@ -5,7 +5,7 @@ function ResponseStatus({ status }: { status: number }) {
   const style = Math.floor(status / 100) === 2 ? 'success' : 'error';
 
   return (
-    status > 0 && (
+    !!status && (
       <Typography component={'p'} className={styles.Status}>
         Status:
         <Typography className={styles[style]} component={'span'}>
